requirements.txt
Flask==2.3.3
kafka-python==2.0.2
Flask-SocketIO==5.3.4
python-dotenv==1.0.0
eventlet==0.33.3